docs : add dependencies swagger,modelMapper (#22)
