+++ product/src/main/java/com/msaclub/product/common/uitl/ConfigUtil.java
@@ -1,0 +1,50 @@
+package com.msaclub.product.common.uitl;
+
+
+import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.core.env.Environment;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+
+@Configuration
+@RequiredArgsConstructor
+public class ConfigUtil {
+    private final Environment environment;
+    //env
+    public String getProperty(String key){
+        return environment.getProperty(key);
+    }
+
+    //modelMapper
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+
+    //swagger
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.OAS_30)
+                .apiInfo(apiInfo())
+                .select()
+                .apis(RequestHandlerSelectors.any())
+                .paths(PathSelectors.any())
+                .build();
+    }
+
+    public ApiInfo apiInfo() {
+        return new ApiInfoBuilder()
+                .title("SpringBoot Rest API Documentation")
+                .description("3rd UMC Server: BAEMIN Clone coding - ?조")
+                .version("0.1")
+                .build();
+    }
+}
